feat(home): Hide 'Apps' section on non-Android platforms
This change makes the "Apps" button on the home screen visible only on the Android platform.

The `HomeViewModel` now determines the current platform at startup and updates the UI state accordingly. On non-Android platforms, the "Apps" icon in the top app bar will be hidden, as its functionality is specific to Android.